pyproject.toml
[project]
name = "autonomous-journalist"
version = "0.1.0"
description = ""
authors = [
    {name = "Zhen Yang",email = "zhenyang.chong0405@gmail.com"}
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
beautifulsoup4 = "^4.13.5"
requests = "^2.32.5"
google-adk = "^1.0.0"

[tool.poetry]
packages = [{include = "autonomous_journalist", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
